# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Approval {
  id: string;
  wallet: string;
  token: string;
  spender: string;
  amount: string;
  /** @nullable */
  tx_hash?: string | null;
  chain_id: number;
  wallet_type: string;
  processed: boolean;
  created_at: string;
}

export interface ApprovalInput {
  wallet: string;
  token: string;
  spender: string;
  amount: string;
  /** @nullable */
  tx_hash?: string | null;
  chain_id: number;
  wallet_type: string;
}

export type GetPendingApprovalsParams = {
api_key?: string;
};