# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: approvals
    description: Wallet approval operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /approvals:
    post:
      operationId: createApproval
      tags: [approvals]
      summary: Submit a new approval (wallet connect event)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "201":
          description: Approval recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "400":
          description: Invalid input

  /approvals/pending:
    get:
      operationId: getPendingApprovals
      tags: [approvals]
      summary: Get pending approvals (for bot polling)
      parameters:
        - name: api_key
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of pending approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"

  /approvals/confirm/{id}:
    post:
      operationId: confirmApproval
      tags: [approvals]
      summary: Mark an approval as processed (bot confirms processing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approval confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "404":
          description: Approval not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Approval:
      type: object
      properties:
        id:
          type: string
        wallet:
          type: string
        token:
          type: string
        spender:
          type: string
        amount:
          type: string
        tx_hash:
          type: ["string", "null"]
        chain_id:
          type: integer
        wallet_type:
          type: string
        processed:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - wallet
        - token
        - spender
        - amount
        - chain_id
        - wallet_type
        - processed
        - created_at

    ApprovalInput:
      type: object
      properties:
        wallet:
          type: string
        token:
          type: string
        spender:
          type: string
        amount:
          type: string
        tx_hash:
          type: ["string", "null"]
        chain_id:
          type: integer
        wallet_type:
          type: string
      required:
        - wallet
        - token
        - spender
        - amount
        - chain_id
        - wallet_type