# lib/api-zod/src/generated/types/approvalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApprovalInput {
  wallet: string;
  token: string;
  spender: string;
  amount: string;
  /** @nullable */
  tx_hash?: string | null;
  chain_id: number;
  wallet_type: string;
}